Fix the bug iterating/deleting entries from the HT.
The issue actually turned out to be the incorrect marking of
remove_entry as a PURE function.  PURE is only for functions which
produce a value; any function which modifies its argument
can/should not be pure.

Oops.

Also managed to break the non-naive based kernel, somehow.  It
seems to never copy any values into the global HT mem...

+++ opt.c
@@ -55,7 +55,7 @@ argparse(int argc, char* argv[])
 		exit(EXIT_FAILURE);
 	}
 	const size_t nbricks = (size_t)bricks[0]*bricks[1]*bricks[2]*n_lod;
-	if(nbricks < ht_size) {
+	if(nbricks < (size_t)ht_size) {
 		g_warning("more hash table entries (%d) than bricks (%zu)!",
 		          ht_size, nbricks);
 	}
